// src/app/shared/models/colors.model.ts
export enum Colors {
  green = '#277C78',
  yellow = '#F2CDAC',
  cyan = '#82C9D7',
  navy = '#626070',
  red = '#C94736',
  purple = '#826CB0',
  turquoise = '#597C7C',
  brown = '#93674F',
  magenta = '#934F6F',
  blue = '#3F82B2',
  grey = '#97A0AC',
  army = '#7F9161',
  pink = '#AF81BA',
  gold = '#CAB361',
  orange = '#BE6C49',
}